Added safety check for room_id<1 in deleteRoom

+++ classes/models/Room.php
@@ -1121,6 +1121,19 @@ class Room
     */
     $room_id = $this->converters->checkRoomId($room_id); // checks room id and converts room id to db room id if necessary (when room hash id was passed)
 
+    $room_id = intval ($room_id);
+    
+    if ($room_id < 1) {
+
+      $returnvalue['success'] = false; // set return value to false
+      $returnvalue['error_code'] = 3; // error code - db error
+      $returnvalue['data'] = false; // returned data
+      $returnvalue['count'] = 0; // returned count of datasets
+
+      return $returnvalue;
+
+    } 
+
     $stmt = $this->db->query('DELETE FROM ' . $this->db->au_rooms . ' WHERE id = :id');
     $this->db->bind(':id', $room_id);
     $err = false;
